Use a Set instead of a BitSet for tracking which types support binary transfer.
A BitSet is a compact representation if we're only considering
builtin types that will have low oids, but if any user defined
types are enabled, all bets are off.  Once the oid counter exceeds
INT_MAX, database connections were failing outright even if no
high oid types used binary transfer because we represent these
oids with negative values that a BitSet cannot handle.

+++ org/postgresql/core/Oid.java
@@ -96,7 +96,7 @@ public class Oid {
 
     public static int valueOf(String oid) throws PSQLException {
         try {
-            return Integer.parseInt(oid);
+            return (int)Long.parseLong(oid);
         } catch (NumberFormatException ex) {
         }
         try {
